change helper func loc
